Consolidate tests into 2 main test files
- Merged test_news_sentiment.py tests into test_ai_analysis_endpoints.py
- Merged test_api_news_sentiment.py tests into test_ai_analysis_endpoints.py
- Removed test_manual_screening.py (utility script, left in tests/scripts/)

Final structure:
- test_data_endpoints.py: All data retrieval endpoint tests (13 tests)
- test_ai_analysis_endpoints.py: All AI analysis endpoint tests + news sentiment tests (48 tests)

Total: 61 passing tests covering all 24 API endpoints
